Moved generator into a generator file
