+++ python/python/ert/ecl/CMakeLists.txt
@@ -27,6 +27,7 @@ set(PYTHON_SOURCES
     ecl_cmp.py
     ecl_sum_var_type.py
     ecl_type.py
+    ecl_grid_generator.py
 )
 
 add_python_package("python.ert.ecl" ${PYTHON_INSTALL_PREFIX}/ert/ecl "${PYTHON_SOURCES}" True)
