+++ python/tests/core/ecl/test_grid.py
@@ -21,7 +21,8 @@ import itertools
 from numpy import linspace
 
 from ert.util import IntVector
-from ert.ecl import EclGrid,EclKW,EclDataType, EclUnitTypeEnum, EclFile
+from ert.ecl import EclGrid, EclKW, EclDataType, EclUnitTypeEnum, EclFile
+from ert.ecl import EclGridGenerator as GridGen
 from ert.ecl.faults import Layer , FaultCollection
 from ert.test import ExtendedTestCase , TestAreaContext
 
@@ -48,21 +49,21 @@ CORNER_HOME = {
 
 def createVolumeTestGridBase(dim, dV, offset=1):
     return [
-            EclGrid.createGrid(dim, dV, offset=0),
-            EclGrid.createGrid(dim, dV, offset=offset),
-            EclGrid.createGrid(dim, dV, offset=offset, irregular_offset=True),
-            EclGrid.createGrid(dim, dV, offset=offset, concave=True),
-            EclGrid.createGrid(dim, dV, offset=offset, irregular=True),
-            EclGrid.createGrid(dim, dV, offset=offset, concave=True, irregular=True),
-            EclGrid.createGrid(dim, dV, offset=offset, irregular_offset=True, concave=True),
-            EclGrid.createGrid(dim, dV, offset=0, faults=True),
-            EclGrid.createGrid(dim, dV, offset=offset, faults=True),
-            EclGrid.createGrid(dim, dV, escape_origo_shift=(100, 100, 0), scale=2),
-            EclGrid.createGrid(dim, dV, escape_origo_shift=(100, 100, 0), scale=0.5),
-            EclGrid.createGrid(dim, dV, escape_origo_shift=(100, 100, 0), translation=(50,50,0)),
-            EclGrid.createGrid(dim, dV, escape_origo_shift=(100, 100, 0), rotate=True),
-            EclGrid.createGrid(dim, dV, escape_origo_shift=(100, 100, 0), misalign=True),
-            EclGrid.createGrid(dim, dV, offset=offset, escape_origo_shift=(100, 100, 0),
+            GridGen.createGrid(dim, dV, offset=0),
+            GridGen.createGrid(dim, dV, offset=offset),
+            GridGen.createGrid(dim, dV, offset=offset, irregular_offset=True),
+            GridGen.createGrid(dim, dV, offset=offset, concave=True),
+            GridGen.createGrid(dim, dV, offset=offset, irregular=True),
+            GridGen.createGrid(dim, dV, offset=offset, concave=True, irregular=True),
+            GridGen.createGrid(dim, dV, offset=offset, irregular_offset=True, concave=True),
+            GridGen.createGrid(dim, dV, offset=0, faults=True),
+            GridGen.createGrid(dim, dV, offset=offset, faults=True),
+            GridGen.createGrid(dim, dV, escape_origo_shift=(100, 100, 0), scale=2),
+            GridGen.createGrid(dim, dV, escape_origo_shift=(100, 100, 0), scale=0.5),
+            GridGen.createGrid(dim, dV, escape_origo_shift=(100, 100, 0), translation=(50,50,0)),
+            GridGen.createGrid(dim, dV, escape_origo_shift=(100, 100, 0), rotate=True),
+            GridGen.createGrid(dim, dV, escape_origo_shift=(100, 100, 0), misalign=True),
+            GridGen.createGrid(dim, dV, offset=offset, escape_origo_shift=(100, 100, 0),
                 irregular_offset=True, concave=True, irregular=True,
                 scale=1.5, translation=(5,5,0), rotate=True,
                 misalign=True)
@@ -70,11 +71,11 @@ def createVolumeTestGridBase(dim, dV, offset=1):
 
 def createContainmentTestBase():
     return [
-            (3,  EclGrid.createGrid((6,6,6), (1,1,1), offset=1)),
-            (10, EclGrid.createGrid((3,3,3), (1,1,1), offset=1, concave=True)),
-            (4,  EclGrid.createGrid((10,10,1), (1,1,1), offset=0., misalign=True)),
+            (3,  GridGen.createGrid((6,6,6), (1,1,1), offset=1)),
+            (10, GridGen.createGrid((3,3,3), (1,1,1), offset=1, concave=True)),
+            (4,  GridGen.createGrid((10,10,1), (1,1,1), offset=0., misalign=True)),
             (3,
-                EclGrid.createGrid((6,6,6), (1,1,1), offset=0.,
+                GridGen.createGrid((6,6,6), (1,1,1), offset=0.,
                     escape_origo_shift=(100, 100, 0),
                     irregular_offset=True, concave=True, irregular=True,
                     scale=1.5, translation=(5,5,0),
@@ -109,7 +110,7 @@ def createWrapperGrid(grid):
                 for i, pos in enumerate(corner_pos)
               ]
 
-    return EclGrid.createSingleCellGrid(corners)
+    return GridGen.createSingleCellGrid(corners)
 
 def average(points):
     p = reduce(lambda a,b: (a[0]+b[0], a[1]+b[1], a[2]+b[2]), points)
@@ -126,14 +127,14 @@ class GridTest(ExtendedTestCase):
         nz = 2000
 
         with self.assertRaises(MemoryError):
-            grid = EclGrid.createRectangular( (nx,ny,nz), (1,1,1))
+            grid = GridGen.createRectangular( (nx,ny,nz), (1,1,1))
 
 
     
     def test_posXYEdge(self):
         nx = 10
         ny = 11
-        grid = EclGrid.createRectangular( (nx,ny,1) , (1,1,1) )
+        grid = GridGen.createRectangular( (nx,ny,1) , (1,1,1) )
         self.assertEqual( grid.findCellCornerXY(0,0,0)  , 0 )
         self.assertEqual( grid.findCellCornerXY(nx,0,0) , nx)
         self.assertEqual( grid.findCellCornerXY(0 , ny , 0) , (nx + 1 ) * ny )
@@ -153,7 +154,7 @@ class GridTest(ExtendedTestCase):
     
     
     def test_dims(self):
-        grid = EclGrid.createRectangular( (10,20,30) , (1,1,1) )
+        grid = GridGen.createRectangular( (10,20,30) , (1,1,1) )
         self.assertEqual( grid.getNX() , 10 )
         self.assertEqual( grid.getNY() , 20 )
         self.assertEqual( grid.getNZ() , 30 )
@@ -165,19 +166,19 @@ class GridTest(ExtendedTestCase):
     
     def test_create(self):
         with self.assertRaises(ValueError):
-            grid = EclGrid.createRectangular( (10,20,30) , (1,1,1) , actnum = [0,1,1,2])
+            grid = GridGen.createRectangular( (10,20,30) , (1,1,1) , actnum = [0,1,1,2])
             
         with self.assertRaises(ValueError):
-            grid = EclGrid.createRectangular( (10,20,30) , (1,1,1) , actnum = IntVector(initial_size = 10))
+            grid = GridGen.createRectangular( (10,20,30) , (1,1,1) , actnum = IntVector(initial_size = 10))
     
         actnum = IntVector(default_value = 1 , initial_size = 6000)
         actnum[0] = 0
         actnum[1] = 0
-        grid = EclGrid.createRectangular( (10,20,30) , (1,1,1) , actnum = actnum)
+        grid = GridGen.createRectangular( (10,20,30) , (1,1,1) , actnum = actnum)
         self.assertEqual( grid.getNumActive( ) , 30*20*10 - 2)
 
     def test_repr_and_name(self):
-        grid = EclGrid.createRectangular((2,2,2), (10,10,10), actnum=[0,0,0,0,1,1,1,1])
+        grid = GridGen.createRectangular((2,2,2), (10,10,10), actnum=[0,0,0,0,1,1,1,1])
         pfx = 'EclGrid('
         rep = repr(grid)
         self.assertEqual(pfx, rep[:len(pfx)])
@@ -192,7 +193,7 @@ class GridTest(ExtendedTestCase):
             self.assertEqual(type(g2.getName()), type(''))
 
     def test_node_pos(self):
-        grid = EclGrid.createRectangular( (10,20,30) , (1,1,1) )
+        grid = GridGen.createRectangular( (10,20,30) , (1,1,1) )
         with self.assertRaises(IndexError):
             grid.getNodePos(-1,0,0)
     
@@ -207,7 +208,7 @@ class GridTest(ExtendedTestCase):
     
     
     def test_truncated_file(self):
-        grid = EclGrid.createRectangular( (10,20,30) , (1,1,1) )
+        grid = GridGen.createRectangular( (10,20,30) , (1,1,1) )
         with TestAreaContext("python/ecl_grid/truncated"):
             grid.save_EGRID( "TEST.EGRID")
     
@@ -222,7 +223,7 @@ class GridTest(ExtendedTestCase):
         nx = 4
         ny = 1
         nz = 1
-        grid = EclGrid.createRectangular( (nx,ny,nz) , (1,1,1) )
+        grid = GridGen.createRectangular( (nx,ny,nz) , (1,1,1) )
         (i,j) = grid.findCellXY( 0.5 , 0.5, 0 )   
         self.assertEqual(i , 0)
         self.assertEqual(j , 0)
@@ -236,7 +237,7 @@ class GridTest(ExtendedTestCase):
         nx = 10
         ny = 23
         nz = 7
-        grid = EclGrid.createRectangular( (nx,ny,nz) , (1,1,1) )
+        grid = GridGen.createRectangular( (nx,ny,nz) , (1,1,1) )
         actnum = grid.exportACTNUM()
         
         self.assertEqual( len(actnum) , nx*ny*nz )
@@ -253,7 +254,7 @@ class GridTest(ExtendedTestCase):
         nx = 10
         ny = 23
         nz = 7
-        grid = EclGrid.createRectangular( (nx,ny,nz) , (1,1,1) )
+        grid = GridGen.createRectangular( (nx,ny,nz) , (1,1,1) )
         with self.assertRaises(IndexError):
             grid.findCellXY( 1 , 1, -1 )   
     
@@ -294,7 +295,7 @@ class GridTest(ExtendedTestCase):
         nx = 10
         ny = 10
         nz = 10
-        grid = EclGrid.createRectangular( (nx,ny,nz) , (1,1,1) )
+        grid = GridGen.createRectangular( (nx,ny,nz) , (1,1,1) )
         kw1 = EclKW("KW" , 1001 , EclDataType.ECL_INT )
         with self.assertRaises(ValueError):
             cp = grid.compressedKWCopy( kw1 )
@@ -304,7 +305,7 @@ class GridTest(ExtendedTestCase):
         nx = 10
         ny = 10
         nz = 10
-        grid = EclGrid.createRectangular( (nx,ny,nz) , (2,3,4) )
+        grid = GridGen.createRectangular( (nx,ny,nz) , (2,3,4) )
     
         (dx,dy,dz) = grid.getCellDims( active_index = 0 )
         self.assertEqual( dx , 2 )
@@ -315,7 +316,7 @@ class GridTest(ExtendedTestCase):
         nx = 10
         ny = 7
         nz = 5
-        grid = EclGrid.createRectangular((nx,ny,nz) , (1,1,1))
+        grid = GridGen.createRectangular((nx,ny,nz) , (1,1,1))
         kw = EclKW( "SWAT" , nx*ny*nz , EclDataType.ECL_FLOAT )
         numpy_3d = grid.create3D( kw )
 
@@ -329,7 +330,7 @@ class GridTest(ExtendedTestCase):
         actnum[2] = 1
         actnum[3] = 1
         
-        grid = EclGrid.createRectangular( (nx,ny,nz) , (1,1,1), actnum = actnum)
+        grid = GridGen.createRectangular( (nx,ny,nz) , (1,1,1), actnum = actnum)
         self.assertEqual( len(grid) , nx*ny*nz )
         self.assertEqual( grid.getNumActive( ) , 4 )
 
@@ -337,7 +338,7 @@ class GridTest(ExtendedTestCase):
     def test_output_units(self):
         n = 10
         a = 1
-        grid = EclGrid.createRectangular( (n,n,n), (a,a,a))
+        grid = GridGen.createRectangular( (n,n,n), (a,a,a))
 
         with TestAreaContext("python/ecl_grid/units"):
             grid.save_EGRID( "CASE.EGRID" , output_unit = EclUnitTypeEnum.ECL_FIELD_UNITS )
@@ -398,7 +399,7 @@ class GridTest(ExtendedTestCase):
     def test_cell_corner_containment(self):
         n = 4
         d = 10
-        grid = EclGrid.createRectangular( (n, n, n), (d, d, d))
+        grid = GridGen.createRectangular( (n, n, n), (d, d, d))
 
         for x, y, z in itertools.product(range(0, n*d+1, d), repeat=3):
             self.assertEqual(
@@ -407,7 +408,7 @@ class GridTest(ExtendedTestCase):
                     )
 
     def test_cell_corner_containment_compatability(self):
-        grid = EclGrid.createRectangular( (3,3,3), (1,1,1) )
+        grid = GridGen.createRectangular( (3,3,3), (1,1,1) )
 
         for x, y, z in itertools.product(range(4), repeat=3):
             for i in range(27):
@@ -420,7 +421,7 @@ class GridTest(ExtendedTestCase):
     def test_cell_face_containment(self):
         n = 4
         d = 10
-        grid = EclGrid.createRectangular( (n, n, n), (d, d, d))
+        grid = GridGen.createRectangular( (n, n, n), (d, d, d))
 
         for x, y, z in itertools.product(range(d/2, n*d, d), repeat=3):
             for axis, direction in itertools.product(range(3), [-1, 1]):
@@ -444,7 +445,7 @@ class GridTest(ExtendedTestCase):
             (20, 20, 20)
             ]
 
-        grid = EclGrid.createSingleCellGrid(points)
+        grid = GridGen.createSingleCellGrid(points)
 
         assertPoint = lambda p : self.assertTrue(
                 grid.cell_contains(p[0], p[1], p[2], 0)
@@ -504,7 +505,7 @@ class GridTest(ExtendedTestCase):
             (25, 25, 25)
             ]
 
-        grid = EclGrid.createSingleCellGrid(points)
+        grid = GridGen.createSingleCellGrid(points)
 
         assertPoint = lambda p : self.assertTrue(
                 grid.cell_contains(p[0], p[1], p[2], 0)
